Added a basic homepage view and URL mapping
In this commit, two files have been modified.

In 'leads/views.py', a new view function 'home_page' has been created, which returns a simple "Hello World!" message as an HTTP response.

In 'setup/urls.py', a URL pattern has been defined to map the root URL ('/') to the 'home_page' view. This means that when users access the root URL of the web application, they will be presented with the "Hello World!" message.