Implemented lead listing on the home page
This commit enhances the 'leads/views.py' and 'templates/home_page.html' files to display a list of leads on the home page of the CRM web application.

In 'leads/views.py':

The view function 'home_page' now retrieves all lead objects from the database using 'Lead.objects.all()' and passes them to the template.
In 'templates/home_page.html':

A simple unordered list (<ul>) is introduced to display the leads.
Using the Django template language, a for loop is utilized to iterate through the list of leads and display each lead's information within list items (<li>).
These changes ensure that when a user visits the home page, they will see a list of all available leads. This is a fundamental step in building the CRM application's user interface.